refactor: expression compilation and evaluation moved to its own type
